+++ src/controls/label.hpp
@@ -23,6 +23,6 @@
             std::string m_text;
             HorizontalAlignment m_horizontal_align = HorizontalAlignment::Left;
             VerticalAlignment m_vertical_align = VerticalAlignment::Center;
-            float m_line_spacing = 5.0f;
+            int m_line_spacing = 5;
     };
 #endif
